Integrate Stripe into the checkout button on cart
Simple stripe checkout button that makes a request
to the back-end in order to retrieve the session
id which it then passes to Stripe to veryfy the
purchase. It then redirects the user to a checkout
page, hosted by Stripe, where the user can input
their card details for payment.
At the moment, both successful and canceled
puchases redirect the user back to the landing
page.

Co-author: Connie Panzariello
<connie.panzariello@gmail.com>

Co-author: Oscar Valenzuela
<oscar.valen115@gmail.com>