Setup persisting session for Stripe redirect
Previously, after the customer got redirected
from Stripe, a new browser session would start,
clearing the React state.
Now, the user token is stored in a browser session
and thus only perists page reloads. After a
reload, the token is recoverd and used to
replenish the React state with the user's user
data and order data.

Co-author: Oscar Valenzuela
<oscar.valen115@gmail.com>