Add order logic post Stripe redirection to client
Refactored promise chain inside the rehydrate
function, whose only job is to reset all the
states after Stripe redirects the customer back
to the client.

If a customer checkout on Stripe's checkout page
is successful, the customer is bneing redirected
to the order history frame and the cart frame if
unsuccessful.

Co-author: Oscar Valenzuela
<oscar.valen115@gmail.com>